test(qa): split causal catalog checks
Punchcard-Session: silver-valley-valley-dt
